\docType{package}
\name{msd16s-package}
\alias{msd16s-package}
\title{Curated dataset of many healthy and moderate-to-severe diarretic gut 16s samples on the 454 FLEX platform.}
\description{
  Data used in 'Diarrhea in young children from low-income countries leads to large-scale alterations in intestinal microbiota composition'. Measurements are number of reads clustered into OTUs (operatonal taxanomic units) by DNAClust.
}
\author{
	Joseph N. Paulson
}
\references{
	Diarrhea in young children from low-income countries leads to large-scale alterations in intestinal microbiota composition.
}

